fix: widen TpDictEntry.block_count from uint16 to uint32
block_count was uint16, capping each term at 65,535 blocks (~8.4M
postings). For corpora >50M docs with common terms, a compacted
segment can exceed this limit. The cast to uint16 silently wraps,
corrupting query results.

The fix is binary-compatible: the old V4 layout used uint16 at bytes
8-9 with a reserved uint16 (always 0) at bytes 10-11. On little-endian
platforms (x86-64, ARM64) this reads identically as uint32. No format
version bump needed; existing indexes work without rebuild.

Changes across 8 files:
- TpDictEntry: uint16 block_count + uint16 reserved -> uint32 block_count
- Remove all (uint16) casts on num_blocks assignments
- Remove all entry.reserved = 0 lines for dict entries
- Widen block_count/block_idx locals in BMW, scan, merge paths
- Widen tp_segment_read_skip_entry block_idx parameter to uint32

V3 segments (TpDictEntryV3) are unchanged -- their uint32 offset limit
makes it impossible for a single term to reach the 65K block cap.

+++ src/segment/scan.c
@@ -23,7 +23,7 @@ void
 tp_segment_read_skip_entry(
 		TpSegmentReader *reader,
 		uint64			 skip_index_offset,
-		uint16			 block_idx,
+		uint32			 block_idx,
 		TpSkipEntry		*skip)
 {
 	uint64 skip_offset;
@@ -436,9 +436,9 @@ tp_segment_posting_iterator_seek(
 		uint32					  target_doc_id,
 		TpSegmentPosting		**posting)
 {
-	uint16		block_count;
+	uint32		block_count;
 	int			left, right, mid;
-	uint16		target_block;
+	uint32		target_block;
 	TpSkipEntry skip;
 
 	if (!iter->initialized || iter->finished)
